fixed input adding bug in variable all pass filter

+++ Audio_1062_DSP/source/effects/components/variable_bandpass_filter.h
@@ -31,7 +31,7 @@ OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 typedef struct {
 	SECONDALLPASSFILTER apf;
-	uint32_t pass_stop;
+	uint8_t pass_stop;
 	float out;
 } VARBANDPASS;
